backend/pyproject.toml
[project]
name = "ai-quiz-backend"
version = "0.3.1"
requires-python = ">=3.10"

dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.27",
  "pydantic>=2.6",
  "python-dotenv>=1.0.0",

  # LangChain core
  "langchain-core>=0.3",

  # GigaChat (НОВЫЙ официальный пакет)
  "langchain-gigachat>=0.1.3",
]

[tool.uvicorn]
factory = false
